+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -544,14 +544,16 @@ module ActiveRecord
         unless lock_id.is_a?(Integer) && lock_id.bit_length <= 63
           raise(ArgumentError, "PostgreSQL requires advisory lock ids to be a signed 64 bit integer")
         end
-        query_value("SELECT pg_try_advisory_lock(#{lock_id})", nil, materialize_transactions: true)
+        ensure_advisory_lock_session!
+        query_value("SELECT pg_try_advisory_lock(#{lock_id})", nil, allow_retry: false, materialize_transactions: true)
       end
 
       def release_advisory_lock(lock_id) # :nodoc:
         unless lock_id.is_a?(Integer) && lock_id.bit_length <= 63
           raise(ArgumentError, "PostgreSQL requires advisory lock ids to be a signed 64 bit integer")
         end
-        query_value("SELECT pg_advisory_unlock(#{lock_id})", nil, materialize_transactions: true)
+        ensure_advisory_lock_session!
+        query_value("SELECT pg_advisory_unlock(#{lock_id})", nil, allow_retry: false, materialize_transactions: true)
       end
 
       def enable_extension(name, **)